refactor(data): share preprocessing per-image helper
